Extract tmux hook helpers
