+++ .triage/phase-AD/findings/AD7-TIMEOUT-001.ttl
@@ -5,7 +5,7 @@
   a triage:Finding ;
   triage:findingId "AD7-TIMEOUT-001" ;
   triage:title "Unbounded tmux subprocess calls with no timeout in local post-send emitter" ;
-  triage:description "AD.7-QA-1 verdict (PR #455, important finding) reports unbounded tmux subprocess calls with no timeout in the local-tmux post-send emitter path. Subprocess calls via Command::new(\"tmux\").output() will hang indefinitely if tmux server becomes unresponsive. Verified via code inspection on all branches s7-s11 that no timeout wrapper exists." ;
+  triage:description "AD.7-QA-1 verdict (PR #455, important finding) reports unbounded tmux subprocess calls with no timeout in the local-tmux post-send emitter path. Subprocess calls via Command::new(\"tmux\").output() will hang indefinitely if tmux server becomes unresponsive. The s11 branch fixed the timeout gap at commit 26c57de5; earlier sprint branches remain open until the standard merge-forward lands on integrate/phase-AD." ;
   triage:phaseId "phase-AD" ;
   triage:triageMode "followup_pass" ;
   triage:category "AD7" ;
@@ -15,8 +15,8 @@
   triage:status "open" ;
   triage:dispatchReady true ;
   triage:triagedAt "2026-07-03T19:45:22Z"^^xsd:dateTime ;
-  triage:highestOpenBranch "feature/pAD-s11-smoke-and-readiness-closeout" ;
-  triage:promoteToBranch "feature/pAD-s11-smoke-and-readiness-closeout" ;
+  triage:highestOpenBranch "feature/pAD-s10-directory-metadata-and-doctor-contract-cleanup" ;
+  triage:promoteToBranch "feature/pAD-s10-directory-metadata-and-doctor-contract-cleanup" ;
   triage:hasOccurrence <urn:atm:triage:occurrence/AD7-TIMEOUT-001/s7/1> ;
   triage:hasOccurrence <urn:atm:triage:occurrence/AD7-TIMEOUT-001/s7/2> ;
   triage:hasOccurrence <urn:atm:triage:occurrence/AD7-TIMEOUT-001/s8/1> ;
@@ -31,7 +31,7 @@
   triage:openOn <urn:atm:triage:worktree/s8/f90f04f6> ;
   triage:openOn <urn:atm:triage:worktree/s9/138e85ea> ;
   triage:openOn <urn:atm:triage:worktree/s10/88941b99> ;
-  triage:openOn <urn:atm:triage:worktree/s11/1dc4f9a0> ;
+  triage:fixedOn <urn:atm:triage:worktree/s11/26c57de5> ;
   triage:absentOn <urn:atm:triage:worktree/s1/2e151d21> ;
   triage:absentOn <urn:atm:triage:worktree/s2/490c9d23> ;
   triage:absentOn <urn:atm:triage:worktree/s3/4ff7cfbc> ;
@@ -132,22 +132,22 @@
   triage:file "crates/atm-core/src/send/hook.rs" ;
   triage:line 562 ;
   triage:snippet "let output = Command::new(\"tmux\").args([\"send-keys\", \"-t\", pane_id.as_str(), \"-l\", message])" ;
-  triage:status "open" ;
-  triage:closed false ;
+  triage:status "closed" ;
+  triage:closed true ;
   triage:branch "feature/pAD-s11-smoke-and-readiness-closeout" ;
-  triage:headSha "1dc4f9a0" ;
-  triage:occursIn <urn:atm:triage:worktree/s11/1dc4f9a0> .
+  triage:headSha "26c57de5" ;
+  triage:occursIn <urn:atm:triage:worktree/s11/26c57de5> .
 
 <urn:atm:triage:occurrence/AD7-TIMEOUT-001/s11/2>
   a triage:Occurrence ;
   triage:file "crates/atm-core/src/send/hook.rs" ;
   triage:line 577 ;
   triage:snippet "let output = Command::new(\"tmux\").args([\"send-keys\", \"-t\", pane_id.as_str(), \"Enter\"])" ;
-  triage:status "open" ;
-  triage:closed false ;
+  triage:status "closed" ;
+  triage:closed true ;
   triage:branch "feature/pAD-s11-smoke-and-readiness-closeout" ;
-  triage:headSha "1dc4f9a0" ;
-  triage:occursIn <urn:atm:triage:worktree/s11/1dc4f9a0> .
+  triage:headSha "26c57de5" ;
+  triage:occursIn <urn:atm:triage:worktree/s11/26c57de5> .
 
 <urn:atm:triage:worktree/s1/2e151d21>
   a triage:WorktreeSnapshot ;
@@ -219,9 +219,9 @@
   triage:headSha "88941b99" ;
   triage:orderIndex 10 .
 
-<urn:atm:triage:worktree/s11/1dc4f9a0>
+<urn:atm:triage:worktree/s11/26c57de5>
   a triage:WorktreeSnapshot ;
   triage:branch "feature/pAD-s11-smoke-and-readiness-closeout" ;
   triage:path "/Users/randlee/Documents/github/atm-core-worktrees/feature/pAD-s11-smoke-and-readiness-closeout" ;
-  triage:headSha "1dc4f9a0" ;
+  triage:headSha "26c57de5" ;
   triage:orderIndex 11 .
